+++ backend/canisters/local_user_index/api/src/lib.rs
@@ -2,8 +2,9 @@ use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
 use types::nns::CryptoAmount;
 use types::{
-    ChannelLatestMessageIndex, ChatId, CommunityId, Cryptocurrency, DiamondMembershipPlanDuration, MessageContent,
-    MessageIndex, PhoneNumber, ReferralType, SuspensionDuration, TimestampMillis, UserId,
+    local_user_index_canister_id, CanisterId, ChannelLatestMessageIndex, ChatId, CommunityId, Cryptocurrency,
+    DiamondMembershipPlanDuration, MessageContent, MessageIndex, PhoneNumber, ReferralType, SuspensionDuration,
+    TimestampMillis, UserId,
 };
 
 mod lifecycle;
@@ -95,17 +96,57 @@ pub struct UserSuspended {
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(from = "UserJoinedGroupPrevious")]
 pub struct UserJoinedGroup {
+    pub user_id: UserId,
+    pub chat_id: ChatId,
+    pub local_user_index_canister_id: CanisterId,
+    pub latest_message_index: Option<MessageIndex>,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct UserJoinedGroupPrevious {
     pub user_id: UserId,
     pub chat_id: ChatId,
     pub latest_message_index: Option<MessageIndex>,
 }
 
+impl From<UserJoinedGroupPrevious> for UserJoinedGroup {
+    fn from(value: UserJoinedGroupPrevious) -> Self {
+        UserJoinedGroup {
+            user_id: value.user_id,
+            chat_id: value.chat_id,
+            local_user_index_canister_id: local_user_index_canister_id(value.chat_id.into()),
+            latest_message_index: value.latest_message_index,
+        }
+    }
+}
+
 #[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(from = "UserJoinedCommunityOrChannelPrevious")]
 pub struct UserJoinedCommunityOrChannel {
     pub user_id: UserId,
     pub community_id: CommunityId,
+    pub local_user_index_canister_id: CanisterId,
     pub channels: Vec<ChannelLatestMessageIndex>,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct UserJoinedCommunityOrChannelPrevious {
+    pub user_id: UserId,
+    pub community_id: CommunityId,
+    pub channels: Vec<ChannelLatestMessageIndex>,
+}
+
+impl From<UserJoinedCommunityOrChannelPrevious> for UserJoinedCommunityOrChannel {
+    fn from(value: UserJoinedCommunityOrChannelPrevious) -> Self {
+        UserJoinedCommunityOrChannel {
+            user_id: value.user_id,
+            community_id: value.community_id,
+            local_user_index_canister_id: local_user_index_canister_id(value.community_id.into()),
+            channels: value.channels,
+        }
+    }
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]