+++ backend/canisters/user/impl/src/model/group_chats.rs
@@ -2,7 +2,7 @@ use crate::model::group_chat::GroupChat;
 use serde::{Deserialize, Serialize};
 use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::HashMap;
-use types::{ChatId, MessageIndex, TimestampMillis, Timestamped};
+use types::{CanisterId, ChatId, MessageIndex, TimestampMillis, Timestamped};
 
 #[derive(Serialize, Deserialize, Default)]
 pub struct GroupChats {
@@ -54,16 +54,22 @@ impl GroupChats {
             || self.removed.last().map(|g| g.timestamp > since).unwrap_or_default()
     }
 
-    pub fn create(&mut self, chat_id: ChatId, now: TimestampMillis) -> bool {
-        self.join(chat_id, None, now);
+    pub fn create(&mut self, chat_id: ChatId, local_user_index_canister_id: CanisterId, now: TimestampMillis) -> bool {
+        self.join(chat_id, local_user_index_canister_id, None, now);
         self.groups_created += 1;
         true
     }
 
-    pub fn join(&mut self, chat_id: ChatId, read_up_to: Option<MessageIndex>, now: TimestampMillis) -> bool {
+    pub fn join(
+        &mut self,
+        chat_id: ChatId,
+        local_user_index_canister_id: CanisterId,
+        read_up_to: Option<MessageIndex>,
+        now: TimestampMillis,
+    ) -> bool {
         match self.group_chats.entry(chat_id) {
             Vacant(e) => {
-                e.insert(GroupChat::new(chat_id, read_up_to, now));
+                e.insert(GroupChat::new(chat_id, local_user_index_canister_id, read_up_to, now));
                 self.removed.retain(|g| g.chat_id != chat_id);
                 true
             }